updated updateJob mutation to drop job and change status
